pyproject.toml
[project]
name = "mqtt-chat"
version = "1.0.0"
description = "A text chat using paho-mqtt"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
